Move EpisodeStats to ACER codes (only used by ACER)
