+++ stable_baselines/a2c/utils.py
@@ -30,61 +30,3 @@ def make_path(path):
     :return: (bool) Whether or not it finished correctly
     """
     return os.makedirs(path, exist_ok=True)
-
-
-class EpisodeStats:
-    def __init__(self, n_steps, n_envs):
-        """
-        Calculates the episode statistics
-
-        :param n_steps: (int) The number of steps to run for each environment
-        :param n_envs: (int) The number of environments
-        """
-        self.episode_rewards = []
-        for _ in range(n_envs):
-            self.episode_rewards.append([])
-        self.len_buffer = deque(maxlen=40)  # rolling buffer for episode lengths
-        self.rewbuffer = deque(maxlen=40)  # rolling buffer for episode rewards
-        self.n_steps = n_steps
-        self.n_envs = n_envs
-
-    def feed(self, rewards, masks):
-        """
-        Update the latest reward and mask
-
-        :param rewards: ([float]) The new rewards for the new step
-        :param masks: ([float]) The new masks for the new step
-        """
-        rewards = np.reshape(rewards, [self.n_envs, self.n_steps])
-        masks = np.reshape(masks, [self.n_envs, self.n_steps])
-        for i in range(0, self.n_envs):
-            for j in range(0, self.n_steps):
-                self.episode_rewards[i].append(rewards[i][j])
-                if masks[i][j]:
-                    reward_length = len(self.episode_rewards[i])
-                    reward_sum = sum(self.episode_rewards[i])
-                    self.len_buffer.append(reward_length)
-                    self.rewbuffer.append(reward_sum)
-                    self.episode_rewards[i] = []
-
-    def mean_length(self):
-        """
-        Returns the average length of each episode
-
-        :return: (float)
-        """
-        if self.len_buffer:
-            return np.mean(self.len_buffer)
-        else:
-            return 0  # on the first params dump, no episodes are finished
-
-    def mean_reward(self):
-        """
-        Returns the average reward of each episode
-
-        :return: (float)
-        """
-        if self.rewbuffer:
-            return np.mean(self.rewbuffer)
-        else:
-            return 0
