+++ stable_baselines/acer/buffer.py
@@ -43,7 +43,7 @@ class Buffer(object):
     def has_atleast(self, frames):
         """
         Check to see if the buffer has at least the asked number of frames
-        
+
         :param frames: (int) The number of frames checked
         :return: (bool) number of frames in buffer >= number asked
         """
@@ -54,7 +54,7 @@ class Buffer(object):
     def can_sample(self):
         """
         Check if the buffer has at least one frame
-        
+
         :return: (bool) if the buffer has at least one frame
         """
         return self.num_in_buffer > 0
@@ -62,7 +62,7 @@ class Buffer(object):
     def decode(self, enc_obs):
         """
         Get the stacked frames of an observation
-        
+
         :param enc_obs: ([float]) the encoded observation
         :return: ([float]) the decoded observation
         """
@@ -89,7 +89,7 @@ class Buffer(object):
     def put(self, enc_obs, actions, rewards, mus, dones, masks):
         """
         Adds a frame to the buffer
-        
+
         :param enc_obs: ([float]) the encoded observation
         :param actions: ([float]) the actions
         :param rewards: ([float]) the rewards
@@ -122,7 +122,7 @@ class Buffer(object):
     def take(self, arr, idx, envx):
         """
         Reads a frame from a list and index for the asked environment ids
-        
+
         :param arr: (np.ndarray) the array that is read
         :param idx: ([int]) the idx that are read
         :param envx: ([int]) the idx for the environments
@@ -137,7 +137,7 @@ class Buffer(object):
     def get(self):
         """
         randomly read a frame from the buffer
-        
+
         :return: ([float], [float], [float], [float], [bool], [float])
                  observations, actions, rewards, mus, dones, maskes
         """
